# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Decision-support API for the Clinician Lab Platform.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CaseStatus = typeof CaseStatus[keyof typeof CaseStatus];


export const CaseStatus = {
  intake: 'intake',
  planning: 'planning',
  fitting: 'fitting',
  review: 'review',
} as const;

export interface Case {
  id: string;
  label: string;
  patientAge: number;
  sex: string;
  anatomicalSite: string;
  missingBodyPart: string;
  race: string;
  retentionMethod: string;
  priorTreatments?: string[];
  /** @nullable */
  ethnicityContext?: string | null;
  status: CaseStatus;
  updatedAt: string;
  /** @nullable */
  reviewDate?: string | null;
}

export interface CaseInput {
  /** @minLength 1 */
  label: string;
  /**
     * @minimum 0
     * @maximum 120
     */
  patientAge: number;
  sex: string;
  anatomicalSite: string;
  missingBodyPart: string;
  race: string;
  retentionMethod: string;
  priorTreatments?: string[];
  ethnicityContext?: string;
}

export type CaseUpdateStatus = typeof CaseUpdateStatus[keyof typeof CaseUpdateStatus];


export const CaseUpdateStatus = {
  intake: 'intake',
  planning: 'planning',
  fitting: 'fitting',
  review: 'review',
} as const;

export interface CaseUpdate {
  /** @minLength 1 */
  label?: string;
  /** @minimum 0 @maximum 120 */
  patientAge?: number;
  sex?: string;
  anatomicalSite?: string;
  missingBodyPart?: string;
  race?: string;
  retentionMethod?: string;
  priorTreatments?: string[];
  ethnicityContext?: string | null;
  status?: CaseUpdateStatus;
  /** @nullable */
  reviewDate?: string | null;
}

export type AnalysisInputUvExposure = typeof AnalysisInputUvExposure[keyof typeof AnalysisInputUvExposure];


export const AnalysisInputUvExposure = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export type AnalysisInputLifestyleDemand = typeof AnalysisInputLifestyleDemand[keyof typeof AnalysisInputLifestyleDemand];


export const AnalysisInputLifestyleDemand = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface AnalysisInput {
  uvExposure: AnalysisInputUvExposure;
  lifestyleDemand: AnalysisInputLifestyleDemand;
  region: string;
  /** @minimum 0 */
  referencePhotoCount?: number;
  /** @minimum 0 */
  waxPatternPhotoCount?: number;
}

export interface SkinToneCombination {
  name: string;
  /**
     * @minItems 3
     * @maxItems 4
     */
  swatches: string[];
  rationale: string;
}

export interface MaterialRecommendation {
  siliconeType: string;
  shoreHardness: string;
  tearStrength: string;
  rationale: string;
}

export interface LongevityForecast {
  estimatedMonths: number;
  nextReviewDate: string;
  rationale: string;
}

export interface CaseAnalysis {
  treatmentOptions: string[];
  /**
     * @minimum 0
     * @maximum 100
     */
  naturalnessScore: number;
  skinToneCombinations: SkinToneCombination[];
  materialRecommendation: MaterialRecommendation;
  fittingGuidance: string[];
  longevity: LongevityForecast;
}

export interface Activity {
  id: string;
  title: string;
  detail: string;
  timestamp: string;
}

export interface Dashboard {
  activeCases: number;
  fittingDue: number;
  reviewDue: number;
  recentActivity: Activity[];
}

export type ClinicianPreferencesManufacturerLimits = {
  shoreRange: string;
  reviewRange: string;
};

export interface ClinicianPreferences {
  defaultShoreHardness: string;
  defaultReviewMonths: number;
  naturalnessPriority: string;
  manufacturerLimits: ClinicianPreferencesManufacturerLimits;
}

export interface PreferencesInput {
  defaultShoreHardness: string;
  /**
     * @minimum 3
     * @maximum 18
     */
  defaultReviewMonths: number;
  naturalnessPriority: string;
}

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;